backend-chp/requirements.txt
django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==11.1.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
